# openapi.yaml
openapi: 3.0.3
info:
  title: Notebook API
  version: 0.0.1
servers:
- url: /
paths:
  /api/contacts:
    get:
      operationId: GetContacts
      parameters:
      - description: "Paging token (opaque string, may be JSON-serialized internally)"
        explode: true
        in: query
        name: pageToken
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - description: Paging
        example: 100
        explode: true
        in: query
        name: pageSize
        required: true
        schema:
          maximum: 1000
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                contacts:
                - name: Егор Щербинин
                  number: "89992226622"
                  email: sherb@mail.ru
                  age: 22
                - name: Никита Дресвянников
                  number: "87772223344"
                  email: dresv234@yamdex.ru
                  age: 22
                pageToken: dsjkfs2h3kj32jd2h3
              schema:
                $ref: "#/components/schemas/GetContacts"
          description: OK
        "400":
          description: Validation error
      summary: Получение всех контактов через пагинацию
    post:
      operationId: PostContact
      requestBody:
        content:
          application/json:
            example:
              name: Иван Петров
              number: "+79991112233"
              age: 25
              city: Москва
              description: Коллега
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
        required: true
      responses:
        "201":
          description: Created
      summary: Создание контакта
  /api/contact/{id}:
    get:
      operationId: GetContact
      parameters:
      - example: 33
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                name: Егор Щербинин
                number: "89992226622"
                email: sherb@mail.ru
                age: 22
              schema:
                $ref: "#/components/schemas/GetContact"
          description: OK
        "404":
          description: Not Found
      summary: Получение конкретного контакта
    patch:
      operationId: PatchContact
      parameters:
      - example: 33
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            example:
              age: 25
              city: Москва
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetContact"
          description: OK
        "404":
          description: Not Found
      summary: Изменение контакта
    put:
      operationId: PutContact
      parameters:
      - example: 33
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            example:
              name: Иван Петров
              number: "+79991112233"
              age: 25
              city: Москва
              description: Коллега
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
        required: true
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      summary: Полное изменение контакта
components:
  schemas:
    GetContacts:
      example:
        pageToken: pageToken
        contacts:
        - number: "82223334455"
          city: Moscow
          name: Alexander
          description: Лучший друг
          age: 11
          email: Alexander22@mail.ru
        - number: "82223334455"
          city: Moscow
          name: Alexander
          description: Лучший друг
          age: 11
          email: Alexander22@mail.ru
      properties:
        contacts:
          items:
            $ref: "#/components/schemas/GetContact"
          minItems: 1
          title: contacts
          type: array
        pageToken:
          nullable: true
          title: pageToken
          type: string
      required:
      - contacts
      title: GetContacts
      type: object
    GetContact:
      example:
        number: "82223334455"
        city: Moscow
        name: Alexander
        description: Лучший друг
        age: 11
        email: Alexander22@mail.ru
      properties:
        name:
          example: Alexander
          nullable: false
          title: name
          type: string
        number:
          example: "82223334455"
          maxLength: 11
          minLength: 11
          nullable: false
          title: number
          type: string
        age:
          example: 11
          maximum: 150
          minimum: 0
          nullable: false
          title: age
          type: integer
        email:
          example: Alexander22@mail.ru
          format: email
          nullable: false
          title: email
          type: string
        city:
          example: Moscow
          nullable: false
          title: city
          type: string
        description:
          example: Лучший друг
          nullable: true
          title: description
          type: string
      required:
      - age
      - name
      - number
      title: GetContact
      type: object
    CreateContactRequest:
      example:
        number: "82223334455"
        city: Moscow
        name: Alexander
        description: Лучший друг
        age: 11
        email: Alexander22@mail.ru
      properties:
        name:
          example: Alexander
          nullable: false
          title: name
          type: string
        number:
          example: "82223334455"
          maxLength: 11
          minLength: 11
          nullable: false
          title: number
          type: string
        age:
          example: 11
          maximum: 150
          minimum: 0
          nullable: false
          title: age
          type: integer
        email:
          example: Alexander22@mail.ru
          format: email
          nullable: false
          title: email
          type: string
        city:
          example: Moscow
          nullable: false
          title: city
          type: string
        description:
          example: Лучший друг
          nullable: true
          title: description
          type: string
      required:
      - age
      - name
      - number
      title: CreateContactRequest
      type: object
    UpdateContactRequest:
      example:
        number: "82223334455"
        city: Moscow
        name: Alexander
        description: Лучший друг
        age: 11
        email: Alexander22@mail.ru
      properties:
        name:
          example: Alexander
          nullable: false
          title: name
          type: string
        number:
          example: "82223334455"
          maxLength: 11
          minLength: 11
          nullable: false
          title: number
          type: string
        age:
          example: 11
          maximum: 150
          minimum: 0
          nullable: false
          title: age
          type: integer
        email:
          example: Alexander22@mail.ru
          format: email
          nullable: false
          title: email
          type: string
        city:
          example: Moscow
          nullable: false
          title: city
          type: string
        description:
          example: Лучший друг
          nullable: true
          title: description
          type: string
      title: UpdateContactRequest
      type: object